requirements.txt
pymupdf
pillow
pytesseract
tqdm